components/application-connector/connector-service/internal/clientcontext/middlewares/clustercontext_test.go
package middlewares

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/kyma-project/kyma/components/application-connector/connector-service/internal/clientcontext"

	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
)

const (
	testTenant = "test-tenant"
	testGroup  = "test-group"
)

func TestClusterContextMiddleware_Middleware(t *testing.T) {

	t.Run("should set empty ctx when no headers specified and cluster ctx not enabled", func(t *testing.T) {
		// given
		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			clusterCtx, ok := ctx.Value(clientcontext.ClusterContextKey).(clientcontext.ClusterContext)
			require.True(t, ok)

			assert.Equal(t, clientcontext.TenantEmpty, clusterCtx.Tenant)
			assert.Equal(t, clientcontext.GroupEmpty, clusterCtx.Group)
			w.WriteHeader(http.StatusOK)
		})

		req, err := http.NewRequest("GET", "/", nil)
		require.NoError(t, err)

		rr := httptest.NewRecorder()

		clusterContextStrategy := clientcontext.NewClusterContextStrategy(false)

		middleware := NewClusterContextMiddleware(clusterContextStrategy)

		// when
		resultHandler := middleware.Middleware(handler)
		resultHandler.ServeHTTP(rr, req)

		// then
		assert.Equal(t, http.StatusOK, rr.Code)
	})

	t.Run("should use header values when cluster ctx enabled", func(t *testing.T) {
		// given
		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			ctx := r.Context()
			clusterCtx, ok := ctx.Value(clientcontext.ClusterContextKey).(clientcontext.ClusterContext)
			require.True(t, ok)

			assert.Equal(t, testTenant, clusterCtx.Tenant)
			assert.Equal(t, testGroup, clusterCtx.Group)
			w.WriteHeader(http.StatusOK)
		})

		req, err := http.NewRequest("GET", "/", nil)
		require.NoError(t, err)
		req.Header.Set(clientcontext.TenantHeader, testTenant)
		req.Header.Set(clientcontext.GroupHeader, testGroup)

		rr := httptest.NewRecorder()

		clusterContextStrategy := clientcontext.NewClusterContextStrategy(true)

		middleware := NewClusterContextMiddleware(clusterContextStrategy)

		// when
		resultHandler := middleware.Middleware(handler)
		resultHandler.ServeHTTP(rr, req)

		// then
		assert.Equal(t, http.StatusOK, rr.Code)
	})

	t.Run("should return 400 if no header provided and cluster context is enabled", func(t *testing.T) {
		// given
		handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			w.WriteHeader(http.StatusOK)
		})

		req, err := http.NewRequest("GET", "/", nil)
		require.NoError(t, err)

		rr := httptest.NewRecorder()

		clusterContextStrategy := clientcontext.NewClusterContextStrategy(true)

		middleware := NewClusterContextMiddleware(clusterContextStrategy)

		// when
		resultHandler := middleware.Middleware(handler)
		resultHandler.ServeHTTP(rr, req)

		// then
		assert.Equal(t, http.StatusBadRequest, rr.Code)
	})
}
